add serach in clients

+++ crm/src/main/java/osfix/ag/crm/domain/WorkControl.java
@@ -32,7 +32,7 @@ public class WorkControl {
     Integer day;
 
     @Column(name = "hours")
-    Long hours;
+    Double hours;
 
     @ManyToOne
     @JsonIgnoreProperties("workControls")
